refactor(components): reorganize analysis components into subfolders
- Move AnalysisHeader, AnalysisTabLayout, AnimatedContainer...
- Update import paths in all affected files
- Create index.ts files in each subfolder for easier imports
- Update main index.ts to include new folder structure
