+++ src/components/github/analysis/components/containers/AnimatedContainer.tsx
