+++ src/components/github/analysis/components/containers/index.ts
@@ -1,0 +1,1 @@
+export { default as AnimatedContainer } from './AnimatedContainer';
