+++ src/components/github/analysis/components/controls/index.ts
@@ -1,0 +1,2 @@
+export { default as HeaderControls } from './HeaderControls';
+export { default as ExpandSwitch } from './ExpandSwitch';
