+++ src/components/github/analysis/components/layout/AnalysisHeader.tsx
@@ -1,6 +1,6 @@
 import { Box, Tooltip, Typography } from "@mui/material";
 import type { ReactNode } from "react";
-import type { ThemeConfig } from "./AnalysisThemes";
+import type { ThemeConfig } from "../AnalysisThemes";
 
 interface AnalysisHeaderProps {
   title: string;
