+++ src/components/github/analysis/components/layout/index.ts
@@ -1,0 +1,3 @@
+export { default as AnalysisTabLayout } from './AnalysisTabLayout';
+export { default as AnalysisHeader } from './AnalysisHeader';
+export { default as HeaderTitle } from './HeaderTitle';
