+++ src/components/github/analysis/components/stats/StatsFooter.tsx
@@ -1,6 +1,6 @@
 import { Box, Typography } from "@mui/material";
 import type { ReactNode } from "react";
-import type { ThemeConfig } from "./AnalysisThemes";
+import type { ThemeConfig } from "../AnalysisThemes";
 
 interface StatsFooterProps {
   title: string;
