+++ src/components/github/analysis/components/stats/index.ts
@@ -1,0 +1,1 @@
+export { default as StatsFooter } from './StatsFooter';
